Cleaned up IO functions
Rewrote assert macros to use our own formatted print instead of fprintf.

Implemented BR_DEBUG which allows the user to print an expression,
its result and return the evaluated expression.

Rewrite numeric limits functions to accept a templated type
so that we can use them in the new print_signed and print_unsigned
functions.

Add functions to print to stderr which involved rewriting the existing
print routines to accept a stream as a parameter.

Add some new traits including:
- is_integral
- is_signed
- is_unsigned