Add some debug macros
Added BR_TODO, BR_DEBUG_RUN and BR_DEBUG macros.

BR_TODO just prints a message when executed to remind you
to do something.

BR_DEBUG_RUN will allow for some expression to be evaluated
in debug mode but not in release mode. Useful for debug logging.

BR_DEBUG prints an expression and its result for debugging purposes.
In release mode, it still evaluates and returns the result of the
expression but doesn't print it.