Add [[nodiscard]] to str_view functions and faster utf_char_length
Added [[nodiscard]] attribute to str_view functions because
we use a value-oriented design so not assigning the result
is most likely a bug.

Simpler utf_char_length function which is more efficient than
the previous iteration.

Added a new is_control classification function to the unicode
utilities for classifying non-visible characters.